fixup! feat(sequel): add plugin :llm
